streamlit_app/requirements.txt
streamlit
streamlit-drawable-canvas
tensorflow
pillow
prefect
pandas
scikit-learn
requests
loguru